+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `default_where` option to add generic bounds on `default`.
+
 ### Changed
 - [**BREAKING**] When `default` is an expression that uses previous fields,
   it'll receive them as reference.